refactor: split code in different files
